Split RSA verification- & signing- specific code into submodules.
